🚚 Rename: pages 폴더에 auth 폴더 생성 후 파일 이동 (#189)
로그인 page, 회원가입 page 이동
