+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import LandingPage from '@/pages/LandingPage'
 import { MyInfo } from '@/pages/my-page/MyInfo'
-import Login from '@/pages/Login'
-import Signup from '@/pages/SignUp'
+import Login from '@/pages/auth/Login'
+import Signup from '@/pages/auth/SignUp'
 import { AppliedRecruitment } from '@/pages/my-page/AppliedRecruitment'
 import Bookmark from '@/pages/my-page/Bookmark'
 import { CompletedStudy } from '@/pages/my-page/CompletedStudy'
